pyproject.toml
[project]
name = "llm-backend"
version = "0.1.0"
description = ""
authors = [{ name = "David Chiu", email = "david20571015@gmail.com" }]
requires-python = "==3.12.*"
readme = "README.md"
dependencies = [
    "grpcio>=1.62.1,<2",
    "grpcio-tools>=1.62.1,<2",
    "llama-index~=0.12.16",
    "llama-index-vector-stores-qdrant~=0.4.3",
    "llama-index-embeddings-huggingface~=0.5.1",
    "llama-index-llms-openai~=0.3.18",
    "qdrant-client~=1.13.2",
    "pydantic>=2.10.6",
]

[project.scripts]
serve = "scripts.serve:main"
gen-protos = "scripts.gen_protos:generate"

[dependency-groups]
dev = ["pre-commit==3.7.1", "ruff==0.9.4"]

[tool.ruff]
extend-exclude = ["llm_backend/protos"]

[tool.ruff.lint]
extend-select = ["I", "UP"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
